페이 기능 개발

+++ src/main/java/comatching/comatching3/match/dto/response/CodeCheckRes.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CodeCheckRes {
-	private Integer currentPoint;
+	private Long currentPoint;
 }
